Refactored forked code.
There are still things to fix.
