+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+	"files.associations": {
+		"*.sqlbook": "sql",
+		"*.ndjson": "jsonl",
+		"*.dbclient-js": "javascript",
+		"stdlib.h": "c",
+		"stdbool.h": "c",
+		"assert.h": "c",
+		"csp.h": "c"
+	}
+}
